backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
werkzeug==2.3.7
requests==2.31.0